Refactor: move engine config and output to separate dir
