+++ openraft/src/engine/handler/leader_handler/mod.rs
@@ -1,10 +1,10 @@
 use std::marker::PhantomData;
 
-use crate::engine::engine_impl::EngineOutput;
 use crate::engine::handler::replication_handler::ReplicationHandler;
 use crate::engine::handler::replication_handler::SendNone;
 use crate::engine::Command;
 use crate::engine::EngineConfig;
+use crate::engine::EngineOutput;
 use crate::entry::RaftEntry;
 use crate::internal_server_state::LeaderQuorumSet;
 use crate::leader::Leader;
