+++ openraft/src/engine/handler/snapshot_handler/mod.rs
@@ -1,4 +1,4 @@
-use crate::engine::engine_impl::EngineOutput;
+use crate::engine::EngineOutput;
 use crate::raft_state::LogStateReader;
 use crate::summary::MessageSummary;
 use crate::Node;
